Code reorganization and optimizations
Commands.py  now holds all the code for the different commands, making
the code neater and easier to modify. USERS.json created as a database
of users on the server(s) containing their userLevel (a ranking of their
permissions) as well as a blacklist flag. That will be expanded in the
future.
Moved testing code into a separate folder.